move validation to account.go
